refactor: return usizes from register methods

+++ parsnips-emu/src/inst/mod.rs
@@ -34,53 +34,53 @@ impl RegFields for Inst {
 }
 
 pub trait ArithLogFields {
-    fn rs(&self) -> u8;
-    fn rt(&self) -> u8;
-    fn rd(&self) -> u8;
+    fn rs(&self) -> usize;
+    fn rt(&self) -> usize;
+    fn rd(&self) -> usize;
 }
 impl ArithLogFields for Inst {
     #[inline(always)]
-    fn rs(&self) -> u8 {
-        (self >> 21 & MASK5) as u8
+    fn rs(&self) -> usize {
+        (self >> 21 & MASK5) as usize
     }
     #[inline(always)]
-    fn rt(&self) -> u8 {
-        (self >> 16 & MASK5) as u8
+    fn rt(&self) -> usize {
+        (self >> 16 & MASK5) as usize
     }
     #[inline(always)]
-    fn rd(&self) -> u8 {
-        (self >> 11 & MASK5) as u8
+    fn rd(&self) -> usize {
+        (self >> 11 & MASK5) as usize
     }
 }
 
 pub trait DivMultFields {
-    fn rs(&self) -> u8;
-    fn rt(&self) -> u8;
+    fn rs(&self) -> usize;
+    fn rt(&self) -> usize;
 }
 impl DivMultFields for Inst {
     #[inline(always)]
-    fn rs(&self) -> u8 {
-        (self >> 21 & MASK5) as u8
+    fn rs(&self) -> usize {
+        (self >> 21 & MASK5) as usize
     }
     #[inline(always)]
-    fn rt(&self) -> u8 {
-        (self >> 16 & MASK5) as u8
+    fn rt(&self) -> usize {
+        (self >> 16 & MASK5) as usize
     }
 }
 
 pub trait ShiftFields {
-    fn rt(&self) -> u8;
-    fn rd(&self) -> u8;
+    fn rt(&self) -> usize;
+    fn rd(&self) -> usize;
     fn shamt(&self) -> u8;
 }
 impl ShiftFields for Inst {
     #[inline(always)]
-    fn rt(&self) -> u8 {
-        (self >> 16 & MASK5) as u8
+    fn rt(&self) -> usize {
+        (self >> 16 & MASK5) as usize
     }
     #[inline(always)]
-    fn rd(&self) -> u8 {
-        (self >> 11 & MASK5) as u8
+    fn rd(&self) -> usize {
+        (self >> 11 & MASK5) as usize
     }
     #[inline(always)]
     fn shamt(&self) -> u8 {
@@ -89,70 +89,70 @@ impl ShiftFields for Inst {
 }
 
 pub trait ShiftVFields {
-    fn rs(&self) -> u8;
-    fn rt(&self) -> u8;
-    fn rd(&self) -> u8;
+    fn rs(&self) -> usize;
+    fn rt(&self) -> usize;
+    fn rd(&self) -> usize;
 }
 impl ShiftVFields for Inst {
     #[inline(always)]
-    fn rs(&self) -> u8 {
-        (self >> 21 & MASK5) as u8
+    fn rs(&self) -> usize {
+        (self >> 21 & MASK5) as usize
     }
     #[inline(always)]
-    fn rt(&self) -> u8 {
-        (self >> 16 & MASK5) as u8
+    fn rt(&self) -> usize {
+        (self >> 16 & MASK5) as usize
     }
     #[inline(always)]
-    fn rd(&self) -> u8 {
-        (self >> 11 & MASK5) as u8
+    fn rd(&self) -> usize {
+        (self >> 11 & MASK5) as usize
     }
 }
 
 pub trait JumpRFields {
-    fn rs(&self) -> u8;
+    fn rs(&self) -> usize;
 }
 impl JumpRFields for Inst {
     #[inline(always)]
-    fn rs(&self) -> u8 {
-        (self >> 21 & MASK5) as u8
+    fn rs(&self) -> usize {
+        (self >> 21 & MASK5) as usize
     }
 }
 
 pub trait MoveFromFields {
-    fn rd(&self) -> u8;
+    fn rd(&self) -> usize;
 }
 impl MoveFromFields for Inst {
     #[inline(always)]
-    fn rd(&self) -> u8 {
-        (self >> 11 & MASK5) as u8
+    fn rd(&self) -> usize {
+        (self >> 11 & MASK5) as usize
     }
 }
 
 pub trait MoveToFields {
-    fn rs(&self) -> u8;
+    fn rs(&self) -> usize;
 }
 impl MoveToFields for Inst {
     #[inline(always)]
-    fn rs(&self) -> u8 {
-        (self >> 21 & MASK5) as u8
+    fn rs(&self) -> usize {
+        (self >> 21 & MASK5) as usize
     }
 }
 
 // immediate encodings
 
 pub trait ArithLogIFields {
-    fn rt(&self) -> u8;
-    fn rs(&self) -> u8;
+    fn rt(&self) -> usize;
+    fn rs(&self) -> usize;
     fn imm(&self) -> u16;
 }
 impl ArithLogIFields for Inst {
     #[inline(always)]
-    fn rs(&self) -> u8 {
-        (self >> 21 & MASK5) as u8
+    fn rs(&self) -> usize {
+        (self >> 21 & MASK5) as usize
     }
     #[inline(always)]
-    fn rt(&self) -> u8 {
-        (self >> 16 & MASK5) as u8
+    fn rt(&self) -> usize {
+        (self >> 16 & MASK5) as usize
     }
     #[inline(always)]
     fn imm(&self) -> u16 {
@@ -161,13 +161,13 @@ impl ArithLogIFields for Inst {
 }
 
 pub trait LoadIFields {
-    fn rt(&self) -> u8;
+    fn rt(&self) -> usize;
     fn imm(&self) -> u16;
 }
 impl LoadIFields for Inst {
     #[inline(always)]
-    fn rt(&self) -> u8 {
-        (self >> 16 & MASK5) as u8
+    fn rt(&self) -> usize {
+        (self >> 16 & MASK5) as usize
     }
     #[inline(always)]
     fn imm(&self) -> u16 {
@@ -176,18 +176,18 @@ impl LoadIFields for Inst {
 }
 
 pub trait BranchFields {
-    fn rs(&self) -> u8;
-    fn rt(&self) -> u8;
+    fn rs(&self) -> usize;
+    fn rt(&self) -> usize;
     fn label(&self, curr: u32) -> u32;
 }
 impl BranchFields for Inst {
     #[inline(always)]
-    fn rs(&self) -> u8 {
-        (self >> 21 & MASK5) as u8
+    fn rs(&self) -> usize {
+        (self >> 21 & MASK5) as usize
     }
     #[inline(always)]
-    fn rt(&self) -> u8 {
-        (self >> 16 & MASK5) as u8
+    fn rt(&self) -> usize {
+        (self >> 16 & MASK5) as usize
     }
     #[inline(always)]
     fn label(&self, curr: u32) -> u32 {
@@ -196,13 +196,13 @@ impl BranchFields for Inst {
 }
 
 pub trait BranchZFields {
-    fn rs(&self) -> u8;
+    fn rs(&self) -> usize;
     fn label(&self, curr: u32) -> u32;
 }
 impl BranchZFields for Inst {
     #[inline(always)]
-    fn rs(&self) -> u8 {
-        (self >> 21 & MASK5) as u8
+    fn rs(&self) -> usize {
+        (self >> 21 & MASK5) as usize
     }
     #[inline(always)]
     fn label(&self, curr: u32) -> u32 {
@@ -211,18 +211,18 @@ impl BranchZFields for Inst {
 }
 
 pub trait LoadStoreFields {
-    fn rt(&self) -> u8;
-    fn rs(&self) -> u8;
+    fn rt(&self) -> usize;
+    fn rs(&self) -> usize;
     fn imm(&self) -> u16;
 }
 impl LoadStoreFields for Inst {
     #[inline(always)]
-    fn rs(&self) -> u8 {
-        (self >> 21 & MASK5) as u8
+    fn rs(&self) -> usize {
+        (self >> 21 & MASK5) as usize
     }
     #[inline(always)]
-    fn rt(&self) -> u8 {
-        (self >> 16 & MASK5) as u8
+    fn rt(&self) -> usize {
+        (self >> 16 & MASK5) as usize
     }
     #[inline(always)]
     fn imm(&self) -> u16 {
